David Winings 4/13/2012 16:43
I redid the level loading code. It will load every .gmap file in the levels
folder in alphabetical order. I still need to make it not crash is the
levels folder is empty, but that should be straightforward and should
also never happen under normal operating conditions. These changes added
dependencies to the python os and re modules, the first for os functions
and the latter for regular expressions (to map the .gmap files).